pyproject.toml
[tool.poetry]
name = "quips"
version = "0.1.0"
description = "Silly web app for silly quips."
authors = ["Brad Smith <bradster@infinitewarp.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
django-environ = "*"
django-memoize = "*"
django-model-utils = "*"
django-redis = "*"
gunicorn = "*"
psycopg2-binary = "*"
pytz = "*"
redis = "*"
unicode-slugify = "*"
whitenoise = "*"
djangorestframework = "*"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
django-extensions = "*"
flake8 = "*"
flake8-docstrings = "*"
flake8-import-order = "*"
flake8-black = "*"
faker = "*"
coverage = "*"
werkzeug = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
